adjust listing form, reuse for editing
